* Changed library view to show directories first * Search filter now properly filters directories. * Fixed typo in EngineFilterBlock * Double->Float in EngineBufferScaleLinear * EngineFilterIIR denormalization applies to OS X now as well (someone think of the Intel machines!) * Use fabs() instead of hack in EngineVUMeter, and tell MSVC to use intrinsic fabs. Most CPUs have hardware fabs these days...

+++ mixxx/src/enginebufferscalelinear.cpp
@@ -63,7 +63,7 @@ void EngineBufferScaleLinear::clear()
 
 CSAMPLE * EngineBufferScaleLinear::scale(double playpos, int buf_size, float * pBase, int iBaseLength)
 {
-    double rate_add = 2.*m_dBaseRate;
+    float rate_add = 2.*m_dBaseRate;
 
     // Determine position in read_buffer to start from
     new_playpos = playpos;
